Split rails and railtie integration
This will allow people to not use the railtie if needed and add the middleware manually
